pyproject.toml
[tool.poetry]
name = "novelist_prototype"
version = "0.1.0"
description = "Interactive Novel Prototype"
authors = ["Hyeon-Jeong, Hwang"]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.24.0"
pytest = "^7.0.0"
python-dotenv = "^1.0.0"
neo4j = "^5.8.1"
langchain-neo4j = "^0.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 